requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.39.0
google-genai>=1.0.0
psutil>=5.9.0
radon>=6.0.0
pylint>=3.0.0
bandit>=1.7.0
scipy>=1.11.0
